requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
lxml==4.9.2
pycryptodome==3.17
requests==2.28.2
urllib3==1.26.15
